refactor: Reorganize code into modular structure (core, calculations, display)
Split monolithic ketu.py (531 lines) into focused modules for better maintainability:

New structure:
- ketu/core.py: Data structures (bodies, aspects, signs)
- ketu/calculations.py: All computation functions (positions, aspects, orbes)
- ketu/display.py: Display and CLI functions
- ketu/ketu.py: Backward compatibility layer (re-exports everything)

Benefits:
- Better code organization (SRP - Single Responsibility Principle)
- Easier to navigate and maintain
- Clear separation of concerns
- Backward compatible (all imports still work)
- 76% code coverage maintained
- All 33 tests pass

No functionality changes, pure refactoring.
